Use a dedicated memory context for executing custom scripts.
This provides better isolation for the execution of custom scripts: the
memory allocated by the parse / plan / execute cycle will be released
early.

The real benefit of this is that the memory manager hooks will be called
once the context is freed. In particular, this helps an extension like
pgaudit which plugs it's events stack lifecycle on the memory context
associated to it. This was found out the hard way when pgaudit started
to complain about having a stack of events even though it was trying to
log a top-level query.

Close #46.